+++ CHANGES.md
@@ -3,8 +3,19 @@ Change Log
 
 ### 1.45 - 2018-05-01
 
+##### Major Announcements :loudspeaker:
+
+* We've launched Cesium ion, read all about it in our [blog post](https://cesium.com/TBD).
+* Cesium now uses ion services by default for base imagery, terrain, and geocoding. [Sign up](https://cesium.com/TBD) for a free ion account to get your own access token.
+
+##### Breaking Changes :mega:
+* `BingMapsImageryProvider` is no longer the default base imagery layer.
+* `BingMapsGeocoderService` is no longer the default geocoder service.
+* Cesium no longer ships with a demo Bing Maps API key.
+
 ##### Additions :tada:
 * Added `IonGeocoderService` and made it the default geocoding service for the `Geocoder` widget.
+* Added `createWorldImagery` which provides Bing Maps imagery via a Cesium ion account.
 * Added option `logarithmicDepthBuffer` to `Scene`. With this option there is typically a single frustum using logarithmic depth rendered. This increases performance by issuing less draw calls to the GPU and helps to avoid artifacts on the connection of two frustums. [#5851](https://github.com/AnalyticalGraphicsInc/cesium/pull/5851)
 * When a log depth buffer is supported, the frustum near and far planes default to `0.1` and `1e10` respectively.
 * Added `Math.log2` to compute the base 2 logarithm of a number.